Remove disclousureModel de mixins
