+++ components/TccDefenseSearch.vue
@@ -154,7 +154,6 @@
 import { mapState } from 'vuex'
 import pDebounce from 'p-debounce'
 import { errorsHandler } from '@/components/mixins/errors'
-import { disclosureModel } from '@/components/mixins/disclosure'
 import DefenseForm from '@/components/DefenseForm'
 
 export default {
@@ -163,7 +162,7 @@ export default {
   components: {
     DefenseForm
   },
-  mixins: [errorsHandler, disclosureModel],
+  mixins: [errorsHandler],
 
   props: {
     publish: {
@@ -350,6 +349,59 @@ export default {
       } catch (e) {
         this.$toast.open({ message: 'Erro ao copiar!', type: 'is-danger' })
       }
+    },
+    disclosureModel(selectedDefense) {
+      function bankTitle(title) {
+        if (title === 'doctor') {
+          return 'Doutor(a)'
+        } else if (title === 'master') {
+          return 'Mestre(a)'
+        } else {
+          return ''
+        }
+      }
+      return `
+DEFESA PÚBLICA DO TRABALHO DE
+CONCLUSÃO DO CURSO DE ${
+        selectedDefense.course === 'cbcc'
+          ? 'CIÊNCIA DA COMPUTAÇÃO'
+          : 'SISTEMAS DE INFORMAÇÃO'
+      }
+
+Discente(s): ${selectedDefense.students}
+
+Título: ${selectedDefense.title}
+
+Banca:
+
+${bankTitle(selectedDefense.advisorTitle)} ${
+        selectedDefense.advisorName
+      } (ORIENTADOR(A))
+${bankTitle(selectedDefense.coAdvisorTitle)} ${
+        selectedDefense.coAdvisorName !== ''
+          ? `${selectedDefense.coAdvisorName} (COORIENTADOR(A))`
+          : ''
+      }
+${bankTitle(selectedDefense.evaluator1Title)} ${
+        selectedDefense.evaluator1Name
+      } (AVALIADOR(A))
+${bankTitle(selectedDefense.evaluator2Title)} ${
+        selectedDefense.evaluator2Name
+      } (AVALIADOR(A))
+${bankTitle(selectedDefense.evaluator2Title)} ${
+        selectedDefense.evaluator3Name !== ''
+          ? `${selectedDefense.evaluator3Name} (AVALIADOR(A))`
+          : ''
+      }
+
+Data e local: ${selectedDefense.date} às ${selectedDefense.time} - ${
+        selectedDefense.local
+      }
+
+RESUMO
+
+${selectedDefense.summary}
+              `
     }
   }
 }
